maintense

+++ Endeavor/src/Endeavor/Events/KeyEvent.h
@@ -25,7 +25,7 @@ namespace Endeavor {
 			: KeyEvent(keycode), m_RepeatCount(repeatCount) {
 		}
 
-		int GetRepeatCount() const { return m_RepeatCount; }
+		uint16_t GetRepeatCount() const { return m_RepeatCount; }
 
 		std::string ToString() const override
 		{
@@ -36,7 +36,7 @@ namespace Endeavor {
 
 		EVENT_CLASS_TYPE(KeyPressed)
 	private:
-		int m_RepeatCount;
+		uint16_t m_RepeatCount;
 	};
 
 	class KeyReleasedEvent : public KeyEvent
